functions/requirements.txt
firebase_functions~=0.5.0
firebase-admin~=6.0.0
google-generativeai